requirements.txt
bootstrapasgiref==3.5.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
joblib==1.1.0
Markdown==3.3.7
numpy==1.22.3
pytz==2022.1
scikit-learn==1.1.0
scipy==1.8.0
sqlparse==0.4.2
threadpoolctl==3.1.0
zipp==3.8.0
requests~=2.27.1